Merge/api

+++ vertnet/service/download.py
@@ -16,15 +16,14 @@ import json
 import logging
 import uuid
 
+DOWNLOAD_VERSION='download.py 2015-08-25T13:09:36+02:00'
+
 SEARCH_CHUNK_SIZE=1000 # limit on documents in a search result: rows per file
 COMPOSE_FILE_LIMIT=32 # limit on the number of files in a single compose request
 COMPOSE_OBJECT_LIMIT=1024 # limit on the number of files in a composition
 TEMP_BUCKET='vn-dltest' # bucket for temp compositions
 DOWNLOAD_BUCKET='vn-downloads2' # production bucket for downloads
 FILE_EXTENSION='tsv'
-
-# def _write_record(f, record):
-#     f.write('%s\n' % record.csv)
 
 def _tsv(json):
     json['datasource_and_rights'] = json.get('url')
@@ -49,7 +48,7 @@ def compose_request(bucketname, filepattern, begin, end):
     begin - the index of the first file in the composition used to grab a range of files.
     end - begin plus the number of files to put in the composition (i.e., end index + 1)
     """
-    
+
     objectlist=[]
     for i in range(begin,end):
         objectdict={}
@@ -89,15 +88,29 @@ class DownloadHandler(webapp2.RequestHandler):
 
     def post(self):
         self.get()
-    
+
     def get(self):
-        count, keywords, email, name, download = map(self.request.get, 
-            ['count', 'keywords', 'email', 'name', 'filepattern'])
-        logging.info(' . '.join([count, keywords, email, name, download]))
+        count, keywords, email, name = map(self.request.get, 
+            ['count', 'keywords', 'email', 'name'])
         q = ' '.join(json.loads(keywords))
-        count = int(count)
+        # count is a limit on the number of records to download
         latlon = self.request.headers.get('X-AppEngine-CityLatLong')
-        if count > SEARCH_CHUNK_SIZE:
+        fromapi = self.request.get('api')
+        
+        if fromapi is not None:
+            # Try to send an indicator to the browser if it came from one.
+            body = 'Downloading results:<br>'
+            body += 'File name: %s<br>' % name
+            body += 'Email: %s<br>' % email
+            body += 'Keywords: %s<br>' % keywords
+            body += 'API: %s<br>' % fromapi
+            logging.info(body)
+            self.response.out.write(body)
+
+        logging.info('API: %s Count: %s Keywords: %s Email: %s Name: %s LatLon: \
+            %s\nVersion: %s' % (fromapi, count, keywords, email, name, latlon, 
+            DOWNLOAD_VERSION) )
+        if count==0 or count > SEARCH_CHUNK_SIZE:
             # The results are larger than SEARCH_CHUNK_SIZE, compose a file for download
             self._queue(q, email, name, latlon)
         else:
@@ -107,34 +120,35 @@ class DownloadHandler(webapp2.RequestHandler):
             self.response.headers['Content-Type'] = "text/tab-separated-values"
             self.response.headers['Content-Disposition'] = "attachment; filename=%s" \
                 % filename
-            records, cursor, count = vnsearch.query(q, count)
-            params = dict(query=q, type='download', count=count, downloader=email, 
-                download=download, latlon=latlon)
-            taskqueue.add(url='/apitracker', params=params, queue_name="apitracker") 
+            records, cursor, count, version = vnsearch.query(q, count)
             data = '%s\n%s' % (util.DWC_HEADER, _get_tsv_chunk(records))
             self.response.out.write(data)
-            
+
             # Write single chunk to file in DOWNLOAD_BUCKET
-            max_retries = 10
+            max_retries = 2
             retry_count = 0
             success = False
+            filepattern = '%s-%s' % (name, uuid.uuid4().hex)
+            filename = '/%s/%s.%s' % (DOWNLOAD_BUCKET, filepattern, 
+                FILE_EXTENSION)
+            logging.info('Writing copy of %s records to  %s.\nVersion: %s' 
+                % (len(records), filename, DOWNLOAD_VERSION) )
             while not success and retry_count < max_retries:
                 try:
-                    filepattern = '%s-%s' % (name, uuid.uuid4().hex)
-                    filename = '/%s/%s.%s' % (DOWNLOAD_BUCKET, filepattern, 
-                        FILE_EXTENSION)
-                    logging.info('download.py: Writing copy of %s records to  %s.' 
-                        % (len(records), filename) )
                     with gcs.open(filename, 'w', content_type='text/tab-separated-values',
                             options={'x-goog-acl': 'public-read'}) as f:
-                        records, cursor, count = vnsearch.query(q, count)
+                        records, cursor, count, version = vnsearch.query(q, count)
                         chunk = '%s\n' % _get_tsv_chunk(records)
                         f.write('%s\n' % util.DWC_HEADER)
                         f.write(chunk)
                         success = True
+                        params = dict(query=q, type='download', count=count, 
+                            downloader=email, download=filename, latlon=latlon)
+                        taskqueue.add(url='/apitracker', params=params, 
+                            queue_name="apitracker") 
                 except Exception, e:
-                    logging.error("download.py: I/O error writing small results to %s." 
-                        % (filename) )
+                    logging.error("I/O error writing small results to \
+                        %s.\nError: %s\nVersion: %s" % (filename,e,DOWNLOAD_VERSION) )
                     retry_count += 1
 #                    raise e
 
@@ -155,31 +169,31 @@ class WriteHandler(webapp2.RequestHandler):
             curs = None
 
         # Write single chunk to file, GCS does not support append
-        max_retries = 10
+        max_retries = 2
         retry_count = 0
         success = False
+
+        records, next_cursor, count, version = \
+            vnsearch.query(q, SEARCH_CHUNK_SIZE, curs=curs)
+        chunk = '%s\n' % _get_tsv_chunk(records)
+        if fileindex==0 and not next_cursor:
+            # This is a query with fewer than SEARCH_CHUNK_SIZE results
+            filename = '/%s/%s.%s' % (TEMP_BUCKET, filepattern, FILE_EXTENSION)
+
         while not success and retry_count < max_retries:
             try:
                 with gcs.open(filename, 'w', content_type='text/tab-separated-values',
                              options={'x-goog-acl': 'public-read'}) as f:
-                    records, next_cursor, count = \
-                        vnsearch.query(q, SEARCH_CHUNK_SIZE, curs=curs)
-                    if not curs:
-                        params = dict(query=q, type='download', count=count, 
-                            downloader=email, download=filename, latlon=latlon)
-                        taskqueue.add(url='/apitracker', params=params, 
-                            queue_name="apitracker") 
-                    chunk = '%s\n' % _get_tsv_chunk(records)
                     if fileindex==0:
                         f.write('%s\n' % util.DWC_HEADER)
                     f.write(chunk)
                     success = True
                     reccount = reccount+len(records)
-                    logging.info('download.py: Download chunk saved to  %s: total %s \
-                        records.' % (filename, reccount) )
+                    logging.info('Download chunk saved to  %s: total %s \
+                        records.\nVersion: %s' % (filename, reccount, DOWNLOAD_VERSION) )
             except Exception, e:
-                logging.error("download.py: I/O error writing chunk to FILE: %s \
-                    for\nQUERY: %s" % (filename,q) )
+                logging.error("I/O error writing chunk to FILE: %s for\nQUERY: %s \
+                    \nError: %s\nVersion: %s" % (filename, q, e, DOWNLOAD_VERSION) )
                 retry_count += 1
 #                raise e
 
@@ -190,7 +204,7 @@ class WriteHandler(webapp2.RequestHandler):
             curs = next_cursor.web_safe_string
         else:
             curs = ''
-        
+
         if curs:
             fileindex = fileindex + 1
             params=dict(q=self.request.get('q'), email=email, name=name, 
@@ -206,7 +220,7 @@ class WriteHandler(webapp2.RequestHandler):
                 # Keep writing search chunks to files
                 taskqueue.add(url='/service/download/write', params=params,
                     queue_name="downloadwrite")
-        
+
         # Finalize and email.
         else:
             params=dict(email=email, name=name, filepattern=filepattern, 
@@ -216,7 +230,8 @@ class WriteHandler(webapp2.RequestHandler):
 
 class ComposeHandler(webapp2.RequestHandler):
     def post(self):
-        email, name, filepattern = map(self.request.get, ['email', 'name', 'filepattern'])
+        q, email, name, filepattern, latlon = map(self.request.get, 
+            ['q', 'email', 'name', 'filepattern', 'latlon'])
         requesttime = self.request.get('requesttime')
         composed_filepattern='%s-cl' % filepattern
         reccount = self.request.get('reccount')
@@ -240,7 +255,7 @@ class ComposeHandler(webapp2.RequestHandler):
         if total_files_to_compose>COMPOSE_FILE_LIMIT:
             # Need to do a composition of compositions
             # Compose first round as sets of COMPOSE_FILE_LIMIT or fewer files
-            
+
             compositions=0
             begin=0
 
@@ -251,8 +266,9 @@ class ComposeHandler(webapp2.RequestHandler):
 
                 composed_filename='%s-%s.%s' % (composed_filepattern, compositions, 
                     FILE_EXTENSION)
-#                logging.info('download.py: Prelim composing files into %s. Begin \
-#                     index: %s End index: %s' % (composed_filename, begin, end) )
+#                logging.info('Prelim composing files into %s. Begin \
+#                     index: %s End index: %s\nVersion: %s' % (composed_filename, \
+#                     begin, end, DOWNLOAD_VERSION) )
                 mbody=compose_request(TEMP_BUCKET, filepattern, begin, end)
                 req = service.objects().compose(
                     destinationBucket=TEMP_BUCKET,
@@ -265,8 +281,8 @@ class ComposeHandler(webapp2.RequestHandler):
 
         composed_filename='%s.%s' % (filepattern,FILE_EXTENSION)
         if compositions==1:
-            logging.info('download.py: %s requires no composition.' 
-                % (composed_filename) )
+            logging.info('%s requires no composition.\nVersion: %s' 
+                % (composed_filename, DOWNLOAD_VERSION) )
         else:
             # Compose remaining files
             fp = filepattern
@@ -274,8 +290,9 @@ class ComposeHandler(webapp2.RequestHandler):
                 # If files were composed, compose them further
                 fp = composed_filepattern
             mbody=compose_request(TEMP_BUCKET, fp, 0, compositions)
-#            logging.info('download.py: Composing %s files into %s\nmbody:\n%s' 
-#                % (compositions,composed_filename, mbody) )
+#            logging.info('Composing %s files into %s\nmbody:\n%s 
+#                \nVersion: %s' % (compositions,composed_filename, mbody, \
+#                DOWNLOAD_VERSION) )
             req = service.objects().compose(
                 destinationBucket=TEMP_BUCKET,
                 destinationObject=composed_filename,
@@ -286,16 +303,21 @@ class ComposeHandler(webapp2.RequestHandler):
         # Now, can we zip the final result?
         # Not directly with GCS. It would have to be done using gsutil in Google 
         # Compute Engine
-        
+
         # Copy the file from temporary storage bucket to the download bucket
         src = '/%s/%s' % (TEMP_BUCKET, composed_filename)
         dest = '/%s/%s' % (DOWNLOAD_BUCKET, composed_filename)
-        gcs.copy2(src, dest)
+        try:
+            gcs.copy2(src, dest)
+        except Exception, e:
+            logging.error("Error copying %s to %s \
+                \nError: %s\nVersion: %s" % (src, dest, e, DOWNLOAD_VERSION) )
 
         mbody=acl_update_request()
         # Change the ACL so that the download file is publicly readable.
-#        logging.info('download.py: Requesting update for /%s/%s\nmbody%s' 
-#            % (DOWNLOAD_BUCKET,composed_filename, mbody) )
+#        logging.info('Requesting update for /%s/%s\nmbody%s \
+#            \nVersion: %s' % (DOWNLOAD_BUCKET,composed_filename, mbody, 
+#            DOWNLOAD_VERSION) )
         req = service.objects().update(
                 bucket=DOWNLOAD_BUCKET,
                 object=composed_filename,
@@ -321,8 +343,11 @@ You can download %s records in the file "%s" within the next 24 hours from
 https://storage.googleapis.com/%s/%s.\nRequest generated: %s\nRequest fulfilled: %s
 """ % (reccount, name, DOWNLOAD_BUCKET, composed_filename, requesttime, resulttime))
 
-        logging.info('download.py: Finalized writing /%s/%s' 
-            % (DOWNLOAD_BUCKET,composed_filename) )
+        params = dict(query=q, type='download', count=reccount, downloader=email, 
+            download=dest, latlon=latlon)
+        taskqueue.add(url='/apitracker', params=params, queue_name="apitracker") 
+        logging.info('Finalized writing /%s/%s\nVersion: %s' 
+            % (DOWNLOAD_BUCKET, composed_filename, DOWNLOAD_VERSION) )
 
         params = dict(filepattern=filepattern, fileindex=total_files_to_compose, 
             compositions=compositions)
@@ -343,32 +368,13 @@ class CleanupHandler(webapp2.RequestHandler):
         # Construct the service object for interacting with the Cloud Storage API.
         service = discovery.build('storage', 'v1', credentials=credentials)
 
-        # Remove all of the chunk files used in the composition.
-        for j in range(total_files_to_compose):
-            filename='%s-%s.%s' % (filepattern, j, FILE_EXTENSION)
-            req = service.objects().delete(bucket=TEMP_BUCKET, object=filename)
-
-            max_retries = 10
-            retry_count = 0
-            success = False
-            # Execute the delete request until successful or until our patience runs out
-            while not success and retry_count < max_retries:
-                try:
-                    resp = req.execute()
-                    success=True
-                except Exception, e:
-                    logging.error("download.py: Error deleting chunk file %s \
-                        attempt %s" % (filename, retry_count+1) )
-                    retry_count += 1
-#                    raise e
-
         if total_files_to_compose>COMPOSE_FILE_LIMIT:
             # Remove the temporary compositions
             for j in range(compositions):
                 filename='%s-%s.%s' % (composed_filepattern, j, FILE_EXTENSION)
                 req = service.objects().delete(bucket=TEMP_BUCKET, object=filename)
 
-                max_retries = 10
+                max_retries = 2
                 retry_count = 0
                 success = False
                 # Try the delete request until successful or until our patience runs out
@@ -377,21 +383,49 @@ class CleanupHandler(webapp2.RequestHandler):
                         resp = req.execute()
                         success=True
                     except Exception, e:
-                        logging.error("download.py: Error deleting composed file \
-                            %s attempt %s" % (filename, retry_count+1) )
+                        logging.error("Error deleting composed file %s \
+                            attempt %s\nError: %s\nVersion: %s" % (filename, 
+                            retry_count+1, e, DOWNLOAD_VERSION) )
                         retry_count += 1
 #                            raise e
 
-        # Delete the temporary composed file from the TEMP_BUCKET
-        req = service.objects().delete(bucket=TEMP_BUCKET, object=composed_filename)
-        resp=req.execute()
-        
-        logging.info('download.py: Finalized cleaning temporary files from /%s' 
-            % (TEMP_BUCKET) )
+        if total_files_to_compose>1:
+            # Remove all of the chunk files used in the composition.
+            for j in range(total_files_to_compose):
+                filename='%s-%s.%s' % (filepattern, j, FILE_EXTENSION)
+                req = service.objects().delete(bucket=TEMP_BUCKET, object=filename)
 
-api = webapp2.WSGIApplication([
+                max_retries = 2
+                retry_count = 0
+                success = False
+                # Execute the delete request until successful or until our patience runs out
+                while not success and retry_count < max_retries:
+                    try:
+                        resp = req.execute()
+                        success=True
+                    except Exception, e:
+                        logging.error("Error deleting chunk file %s attempt %s\nError: \
+                            %s\nVersion: %s" % (filename, retry_count+1, e, 
+                            DOWNLOAD_VERSION) )
+                        retry_count += 1
+#                        raise e
+
+            # Delete the temporary composed file from the TEMP_BUCKET
+            req = service.objects().delete(bucket=TEMP_BUCKET, object=composed_filename)
+            try:
+                resp=req.execute()
+            except Exception, e:
+                logging.error("Error deleting temporary composed file %s \
+                    \nError: %s\nVersion: %s" % (filename, e, DOWNLOAD_VERSION) )
+
+            logging.info('Finalized cleaning temporary files from /%s\nVersion: %s' 
+                % (TEMP_BUCKET, DOWNLOAD_VERSION) )
+
+routes = [
     webapp2.Route(r'/service/download', handler=DownloadHandler),
     webapp2.Route(r'/service/download/write', handler=WriteHandler),
     webapp2.Route(r'/service/download/compose', handler=ComposeHandler),
     webapp2.Route(r'/service/download/cleanup', handler=CleanupHandler)
-    ], debug=True)
+    ]
+    
+api = webapp2.WSGIApplication(routes, debug=True)
